requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
python-multipart==0.0.26
python-dotenv==1.2.2
requests==2.33.1
imageio-ffmpeg==0.6.0
g2p-en==2.1.0
nltk==3.9.4

numpy==1.26.4
joblib==1.5.3
protobuf==5.29.5

tensorflow==2.19.0
tf-keras==2.19.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu

transformers==4.48.3
tweet-preprocessor==0.6.0
deepface==0.0.93
google-generativeai==0.8.4
openai-whisper==20250625

av==17.0.0
mediapipe==0.10.33
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
facenet-pytorch==2.6.0
fvcore==0.1.5.post20221221
librosa==0.11.0
moviepy==1.0.3
